Move command handler decorators to decorator folder.
